ble_ws_service/requirements.txt
bleak
yalexs-ble
websockets
PyYAML